refactor: modularized handlers
